Some more pylint things
